Recover index: index.tsx를 원래대로 복구하고 input.tsx파일 생성
